src/app/requirements.txt
Flask~=3.1.1
gunicorn~=23.0.0
ouilookup~=0.3.1
RouterOS-api~=0.21.0
pydantic~=2.11.7
PyYAML~=6.0.2